requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
attrs==19.1.0
Automat==0.7.0
bonjour==0.0.3
cachetools==3.1.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
constantly==15.1.0
cryptography==2.6.1
Flask==1.0.2
Flask-Cors==3.0.7
google-api-core==1.11.0
google-auth==1.6.3
google-cloud-tasks==1.1.0
googleapis-common-protos==1.5.10
grpc-google-iam-v1==0.11.4
grpcio==1.20.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
isort==4.3.17
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
PyHamcrest==1.9.0
pyOpenSSL==19.0.0
pytz==2019.1
redis==3.2.1
requests==2.21.0
rsa==4.0
sentry-sdk==0.7.14
service-identity==18.1.0
simple-salesforce==0.74.2
six==1.12.0
Twisted==19.2.0
typed-ast==1.3.4
urllib3==1.24.2
Werkzeug==0.15.2
wrapt==1.11.1
zope.interface==4.6.0
